refactor: stream descriptor tag
